// apps/account-service/src/infrastructure/types.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Sessions {
  created_at: Generated<Timestamp>;
  expires_at: Timestamp;
  ip: string | null;
  revoked: Generated<boolean>;
  token: string;
  user_agent: string | null;
  user_id: number;
  uuid: string;
}

export interface Users {
  avatar: string | null;
  created_at: Generated<Timestamp>;
  email: string;
  id: Generated<number>;
  password_hash: string;
  screen_name: string | null;
  updated_at: Generated<Timestamp>;
}

export interface DB {
  sessions: Sessions;
  users: Users;
}